fix(backup): consolidate data-mover image and pod defaults helpers
Move data_mover_image(), hardened_security_context(),
hardened_pod_security_context(), and default_resource_requirements()
from libs/backup to libs/backup-core for reuse by the kanidm
operator's transport sidecar builder.

- Create libs/backup-core/src/image.rs with data_mover_image()
- Create libs/backup-core/src/pod_defaults.rs with security contexts
  and resource requirements
- Update libs/backup/src/controller/mod.rs to re-export from backup-core
- Update libs/operator/src/kanidm/restore.rs to use consolidated
  data_mover_image()

This eliminates duplication and prepares for the transport sidecar
injection in the kanidm StatefulSet.
