+++ libs/backup-core/src/lib.rs
@@ -1,7 +1,9 @@
 pub mod auth;
 pub mod crd;
+pub mod image;
 pub mod manifest;
 pub mod operation;
 pub mod paths;
+pub mod pod_defaults;
 pub mod result;
 pub mod retention;
