+++ libs/backup/src/controller/mod.rs
@@ -8,11 +8,13 @@ pub use discovery::CONTROLLER_ID as DISCOVERY_CONTROLLER_ID;
 pub use repository::CONTROLLER_ID as REPOSITORY_CONTROLLER_ID;
 pub use schedule::CONTROLLER_ID as SCHEDULE_CONTROLLER_ID;
 
-use k8s_openapi::api::core::v1::{
-    Capabilities, Pod, PodSecurityContext, ResourceRequirements, SeccompProfile, SecurityContext,
-};
-use k8s_openapi::apimachinery::pkg::api::resource::Quantity;
+use k8s_openapi::api::core::v1::Pod;
 use kube::api::PropagationPolicy;
+
+pub use kaniop_backup_core::image::data_mover_image;
+pub use kaniop_backup_core::pod_defaults::{
+    default_resource_requirements, hardened_pod_security_context, hardened_security_context,
+};
 
 pub const RESULT_PATH: &str = "/kaniop-result/result.json";
 pub const TERMINATION_MESSAGE_LIMIT: usize = 4096;
@@ -21,58 +23,6 @@ pub const BACKUP_JOB_TTL_SECONDS: i32 = 600;
 pub fn background_delete_params() -> kube::api::DeleteParams {
     kube::api::DeleteParams {
         propagation_policy: Some(PropagationPolicy::Background),
-        ..Default::default()
-    }
-}
-
-pub fn data_mover_image() -> String {
-    std::env::var("DATA_MOVER_IMAGE")
-        .unwrap_or_else(|_| "ghcr.io/pando85/kaniop-data-mover:latest".to_string())
-}
-
-pub fn hardened_security_context() -> SecurityContext {
-    SecurityContext {
-        allow_privilege_escalation: Some(false),
-        capabilities: Some(Capabilities {
-            drop: Some(vec!["ALL".to_string()]),
-            ..Default::default()
-        }),
-        read_only_root_filesystem: Some(true),
-        run_as_non_root: Some(true),
-        run_as_user: Some(65534),
-        ..Default::default()
-    }
-}
-
-pub fn hardened_pod_security_context() -> PodSecurityContext {
-    PodSecurityContext {
-        run_as_non_root: Some(true),
-        seccomp_profile: Some(SeccompProfile {
-            type_: "RuntimeDefault".to_string(),
-            ..Default::default()
-        }),
-        ..Default::default()
-    }
-}
-
-pub fn default_resource_requirements() -> ResourceRequirements {
-    ResourceRequirements {
-        requests: Some(
-            [
-                ("cpu".to_string(), Quantity("50m".to_string())),
-                ("memory".to_string(), Quantity("32Mi".to_string())),
-            ]
-            .into_iter()
-            .collect(),
-        ),
-        limits: Some(
-            [
-                ("cpu".to_string(), Quantity("100m".to_string())),
-                ("memory".to_string(), Quantity("64Mi".to_string())),
-            ]
-            .into_iter()
-            .collect(),
-        ),
         ..Default::default()
     }
 }
